# swagger/definitions/swagger.yml
openapi: "3.0.0"
info:
  title: Companies Service
  version: 1.0.0
  description: Companies Service API documentation
servers:
  - url: http://localhost:5002
  - url: https://stage.onelike.gusisoft.cl
tags:
  - name: Companies
    description: Companies Operations
  - name: Classifications
    description: Companies Classifications
  - name: Service Health
    description: Service Health Status
paths:
  /companies:
    get:
      tags:
        - Companies
      summary: List Companies
      description: List Companies by user Token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Companies List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#components/schemas/Company"
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Unauthorized"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
    post:
      tags:
        - Companies
      summary: Create a Company
      description: Create a Company
      security:
        - bearerAuth: []
      requestBody:
        description: Company data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/CompanyCreate'
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#components/schemas/Company"
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#components/responses/BadRequest"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
  /companies/{id}:
    get:
      tags:
        - Companies
      summary: Get Company
      description: Get a Company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company Get
          content:
            application/json:
              schema:
                $ref: "#components/schemas/Company"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
        '404':
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
    put:
      tags:
        - Companies
      summary: Update Company
      description: Update a Company
      security:
        - bearerAuth: []
      requestBody:
        description: Company data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/CompanyCreate'
      parameters:
        - name: id
          in: path
          description: Company id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company Updated
          content:
            application/json:
              schema:
                $ref: "#components/schemas/Company"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
        '404':
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
  /companies/{id}/deactivate:
    put:
      tags:
        - Companies
      summary: Deactivate Company
      description: Deactivate an existing Company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company successfully deactivated.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/Company"
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Unauthorized"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
  /companies/{id}/activate:
    put:
      tags:
        - Companies
      summary: Activate Company
      description: Activate an existing Company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company successfully activated.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/Company"
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Unauthorized"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
  /companies/{id}/users:
    get:
      tags:
        - Companies
      summary: Get Company users
      description: Get Company users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#components/schemas/User"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
        '404':
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
  /companies/users:
    get:
      tags:
        - Companies
      summary: Get Company users
      description: Get Company users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Company id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#components/schemas/User"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
        '404':
          description: Company not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
  /companies/classifications:
    get:
      tags:
        - Classifications
      summary: List Classifications
      description: List Classifications for companies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Classification List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#components/schemas/Classification"
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#components/responses/Forbidden"
  /companies-service/health:
    get:
      tags:
        - Service Health
      summary: Service health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
components:
  schemas:
    Classification:
      properties:
        id:
          type: integer
          format: int64
          description: Classification id
        name:
          type: string
          description: Classification name
    Company:
      allOf:
        - $ref: "#/components/requestBodies/CompanyBase"
        - optional:
          properties:
            id:
              type: integer
              format: int64
              description: Company id
            classification:
              type: string
              description: Company Classification
            created:
              type: string
              format: date-time
              description: Created datetime
            created_by:
              type: integer
              description: User id who creates this
            updated:
              type: string
              format: date-time
              description: Updated datetime
            updated_by:
              type: integer
              description: User id who updates this
    User:
      properties:
        id:
          type: integer
          format: int64
          description: User id
        first_name:
          type: string
          description: User first name
        last_name:
          type: string
          description: User last name
        email:
          type: string
          description: User email
    Health:
      properties:
        message:
          type: string
          description: Health status
  requestBodies:
    CompanyCreate:
      allOf:
        - $ref: "#/components/requestBodies/CompanyBase"
        - required:
          - name
          - classification_id
          - identifier
          properties:
            classification_id:
              type: integer
              format: int64
              description: Classification id
    CompanyBase:
      properties:
        identifier:
          type: string
          description: Company TAX identifier
        name:
          type: string
          description: Company name
        expiration:
          type: string
          format: date-time
          description: Expiration datetime
  responses:
    BadRequest:
      properties:
        message:
          type: string
          description: Error cause
    NotFound:
      properties:
        message:
          type: string
          description: Error cause
    Forbidden:
      properties:
        message:
          type: string
          description: Error cause
    Unauthorized:
      properties:
        message:
          type: string
          description: Error cause

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer